@@ -16,64 +16,47 @@ limitations under the License.
 
 package controllers
 
-import (
-	"path/filepath"
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-
-	infrav1 "github.com/kubesphere/kubekey/api/v1beta1"
-	//+kubebuilder:scaffold:imports
-)
-
-// These tests use Ginkgo (BDD-style Go testing framework). Refer to
-// http://onsi.github.io/ginkgo/ to learn more about Ginkgo.
-
-// var cfg *rest.Config
-var k8sClient client.Client
-var testEnv *envtest.Environment
-
-func TestAPIs(t *testing.T) {
-	RegisterFailHandler(Fail)
-
-	RunSpecsWithDefaultAndCustomReporters(t,
-		"Controller Suite",
-		[]Reporter{printer.NewlineReporter{}})
-}
-
-var _ = BeforeSuite(func() {
-	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
-
-	By("bootstrapping test environment")
-	testEnv = &envtest.Environment{
-		CRDDirectoryPaths:     []string{filepath.Join("..", "config", "crd", "bases")},
-		ErrorIfCRDPathMissing: true,
-	}
-
-	cfg, err := testEnv.Start()
-	Expect(err).NotTo(HaveOccurred())
-	Expect(cfg).NotTo(BeNil())
-
-	err = infrav1.AddToScheme(scheme.Scheme)
-	Expect(err).NotTo(HaveOccurred())
-
-	//+kubebuilder:scaffold:scheme
-
-	k8sClient, err = client.New(cfg, client.Options{Scheme: scheme.Scheme})
-	Expect(err).NotTo(HaveOccurred())
-	Expect(k8sClient).NotTo(BeNil())
-
-}, 60)
-
-var _ = AfterSuite(func() {
-	By("tearing down the test environment")
-	err := testEnv.Stop()
-	Expect(err).NotTo(HaveOccurred())
-})
+//// These tests use Ginkgo (BDD-style Go testing framework). Refer to
+//// http://onsi.github.io/ginkgo/ to learn more about Ginkgo.
+//
+//// var cfg *rest.Config
+//var k8sClient client.Client
+//var testEnv *envtest.Environment
+//
+//func TestAPIs(t *testing.T) {
+//	RegisterFailHandler(Fail)
+//
+//	RunSpecsWithDefaultAndCustomReporters(t,
+//		"Controller Suite",
+//		[]Reporter{printer.NewlineReporter{}})
+//}
+//
+//var _ = BeforeSuite(func() {
+//	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
+//
+//	By("bootstrapping test environment")
+//	testEnv = &envtest.Environment{
+//		CRDDirectoryPaths:     []string{filepath.Join("..", "config", "crd", "bases")},
+//		ErrorIfCRDPathMissing: true,
+//	}
+//
+//	cfg, err := testEnv.Start()
+//	Expect(err).NotTo(HaveOccurred())
+//	Expect(cfg).NotTo(BeNil())
+//
+//	err = infrav1.AddToScheme(scheme.Scheme)
+//	Expect(err).NotTo(HaveOccurred())
+//
+//	//+kubebuilder:scaffold:scheme
+//
+//	k8sClient, err = client.New(cfg, client.Options{Scheme: scheme.Scheme})
+//	Expect(err).NotTo(HaveOccurred())
+//	Expect(k8sClient).NotTo(BeNil())
+//
+//}, 60)
+//
+//var _ = AfterSuite(func() {
+//	By("tearing down the test environment")
+//	err := testEnv.Stop()
+//	Expect(err).NotTo(HaveOccurred())
+//})
